# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Agent Sandbox World",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/actions": {
      "post": {
        "summary": "Act",
        "operationId": "act_api_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Action"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/incidents/runs": {
      "post": {
        "summary": "Start Incident",
        "operationId": "start_incident_api_incidents_runs_post",
        "parameters": [
          {
            "name": "x-lab-capability",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lab-Capability"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/runs/{run_id}": {
      "get": {
        "summary": "Observe Incident",
        "operationId": "observe_incident_api_incidents_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-lab-capability",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lab-Capability"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/runs/{run_id}/tools": {
      "post": {
        "summary": "Incident Tool",
        "operationId": "incident_tool_api_incidents_runs__run_id__tools_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-lab-capability",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lab-Capability"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentToolResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/runs/{run_id}/proposals/validate": {
      "post": {
        "summary": "Validate Proposal",
        "operationId": "validate_proposal_api_incidents_runs__run_id__proposals_validate_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-lab-capability",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lab-Capability"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryProposal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/runs/{run_id}/proposals": {
      "post": {
        "summary": "Request Execution",
        "operationId": "request_execution_api_incidents_runs__run_id__proposals_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-lab-capability",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lab-Capability"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/runs/{run_id}/approval": {
      "post": {
        "summary": "Approve Execution",
        "operationId": "approve_execution_api_incidents_runs__run_id__approval_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-lab-capability",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lab-Capability"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/runs/{run_id}/apply": {
      "post": {
        "summary": "Apply Execution",
        "operationId": "apply_execution_api_incidents_runs__run_id__apply_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-lab-capability",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lab-Capability"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Action": {
        "properties": {
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "actor_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Actor Id"
          },
          "type": {
            "type": "string",
            "const": "move",
            "title": "Type"
          },
          "dx": {
            "type": "integer",
            "title": "Dx"
          },
          "dy": {
            "type": "integer",
            "title": "Dy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["action_id", "actor_id", "type", "dx", "dy"],
        "title": "Action"
      },
      "ActionResult": {
        "properties": {
          "world": {
            "$ref": "#/components/schemas/WorldState"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "type": "array",
            "maxItems": 80,
            "title": "Events"
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["world", "events", "event"],
        "title": "ActionResult"
      },
      "ApplyRequest": {
        "properties": {
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "proposal_hash": {
            "type": "string",
            "title": "Proposal Hash"
          },
          "expected_revision": {
            "type": "integer",
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["action_id", "proposal_hash", "expected_revision"],
        "title": "ApplyRequest"
      },
      "ApprovalRequest": {
        "properties": {
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "proposal_hash": {
            "type": "string",
            "title": "Proposal Hash"
          },
          "expected_revision": {
            "type": "integer",
            "title": "Expected Revision"
          },
          "approve": {
            "type": "boolean",
            "title": "Approve"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["action_id", "proposal_hash", "expected_revision", "approve"],
        "title": "ApprovalRequest"
      },
      "Entity": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "position": {
            "$ref": "#/components/schemas/Position"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "position"],
        "title": "Entity"
      },
      "Event": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "status": {
            "type": "string",
            "enum": ["success", "failure"],
            "title": "Status"
          },
          "reason": {
            "type": "string",
            "enum": ["moved", "out_of_bounds", "unknown_actor", "invalid_move"],
            "title": "Reason"
          },
          "before": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Position"
              },
              {
                "type": "null"
              }
            ]
          },
          "after": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Position"
              },
              {
                "type": "null"
              }
            ]
          },
          "world_id": {
            "type": "string",
            "format": "uuid",
            "title": "World Id"
          },
          "world_revision": {
            "type": "integer",
            "title": "World Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "action",
          "status",
          "reason",
          "before",
          "after",
          "world_id",
          "world_revision"
        ],
        "title": "Event"
      },
...
      "EvidenceItem": {
        "properties": {
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["ref", "source", "timestamp", "data"],
        "title": "EvidenceItem"
      },
      "ExecutionRequest": {
        "properties": {
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "proposal": {
            "$ref": "#/components/schemas/RecoveryProposal"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["action_id", "proposal"],
        "title": "ExecutionRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidentRunRequest": {
        "properties": {
          "scenario_number": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Scenario Number"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["scenario_number"],
        "title": "IncidentRunRequest"
      },
      "IncidentSnapshot": {
        "properties": {
          "scenario_number": {
            "type": "integer",
            "title": "Scenario Number"
          },
          "summary": {
            "$ref": "#/components/schemas/IncidentSummary"
          },
          "services": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Services"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "pending_proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecoveryProposal"
              },
              {
                "type": "null"
              }
            ]
          },
          "proposal_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal Hash"
          },
          "approved_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Hash"
          },
          "applied_changes": {
            "items": {
              "$ref": "#/components/schemas/RecoveryChange"
            },
            "type": "array",
            "title": "Applied Changes",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["scenario_number", "summary", "services", "evidence"],
        "title": "IncidentSnapshot"
      },
      "IncidentSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "incomplete": {
            "type": "integer",
            "title": "Incomplete"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "duplicate_count": {
            "type": "integer",
            "title": "Duplicate Count"
          },
          "state": {
            "type": "string",
            "enum": [
              "investigating",
              "waiting_approval",
              "applying",
              "verifying",
              "recovered",
              "stopped"
            ],
            "title": "State"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "revision",
          "accepted",
          "incomplete",
          "completed",
          "duplicate_count",
          "state"
        ],
        "title": "IncidentSummary"
      },
      "IncidentToolRequest": {
        "properties": {
          "tool": {
            "type": "string",
            "enum": [
              "observe_system",
              "query_events",
              "inspect_queue",
              "read_service_config",
              "list_changes",
              "search_knowledge",
              "probe_dependency",
              "advance_and_verify",
              "lookup_action_status"
            ],
            "title": "Tool"
          },
          "args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Args"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["tool"],
        "title": "IncidentToolRequest"
      },
      "IncidentToolResult": {
        "properties": {
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "status": {
            "type": "string",
            "enum": ["success", "failure", "denied"],
            "title": "Status"
          },
          "summary": {
            "$ref": "#/components/schemas/IncidentSummary"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Evidence"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["tool", "status", "summary", "evidence"],
        "title": "IncidentToolResult"
      },
      "Position": {
        "properties": {
          "x": {
            "type": "integer",
            "title": "X"
          },
          "y": {
            "type": "integer",
            "title": "Y"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["x", "y"],
        "title": "Position"
      },
      "ProposalResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "valid",
              "invalid",
              "waiting_approval",
              "approved",
              "applied",
              "failure",
              "stale"
            ],
            "title": "Status"
          },
          "proposal_hash": {
            "type": "string",
            "title": "Proposal Hash"
          },
          "summary": {
            "$ref": "#/components/schemas/IncidentSummary"
          },
          "violations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Violations",
            "default": []
          },
          "event_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Event Refs",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["status", "proposal_hash", "summary"],
        "title": "ProposalResult"
      },
      "RecoveryChange": {
        "properties": {
          "operation": {
            "type": "string",
            "enum": ["set_config", "select_deployment", "requeue"],
            "title": "Operation"
          },
          "service": {
            "type": "string",
            "enum": ["worker", "carrier", "queue"],
            "title": "Service"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "order_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Order Ids",
            "default": []
          },
          "expected_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected State"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["operation", "service"],
        "title": "RecoveryChange"
      },
      "RecoveryProposal": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "title": "Expected Revision"
          },
          "diagnosis": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Diagnosis"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Evidence Refs"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/RecoveryChange"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Changes"
          },
          "rollback": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Rollback"
          },
          "verification_ticks": {
            "type": "integer",
            "const": 4,
            "title": "Verification Ticks",
            "default": 4
          },
          "canary_count": {
            "type": "integer",
            "const": 2,
            "title": "Canary Count",
            "default": 2
          },
          "verification_scope": {
            "type": "string",
            "const": "affected_orders_and_canaries",
            "title": "Verification Scope",
            "default": "affected_orders_and_canaries"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["expected_revision", "diagnosis", "evidence_refs", "changes", "rollback"],
        "title": "RecoveryProposal"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "WorldState": {
        "properties": {
          "world_id": {
            "type": "string",
            "format": "uuid",
            "title": "World Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "type": "array",
            "title": "Entities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["world_id", "revision", "width", "height", "entities"],
        "title": "WorldState"
      }
    }
  }
}